--- a/src/main/scala/chiseltest/DecoupledDriver.scala
+++ b/src/main/scala/chiseltest/DecoupledDriver.scala
@@ -14,19 +14,6 @@ class DecoupledDriver[T <: Data](x: ReadyValidIO[T]) {
     this
   }
 
-  def setSourceClock(clock: Clock): this.type = {
-    ClockResolutionUtils.setClock(DecoupledDriver.decoupledSourceKey, x, clock)
-    this
-  }
-
-  protected def getSourceClock: Clock = {
-    ClockResolutionUtils.getClock(
-      DecoupledDriver.decoupledSourceKey,
-      x,
-      x.ready.getSourceClock()
-    ) // TODO: validate against bits/valid sink clocks
-  }
-
   def enqueueNow(data: T): Unit = {
     // TODO: check for init
     x.bits.poke(data)
@@ -35,7 +22,7 @@ class DecoupledDriver[T <: Data](x: ReadyValidIO[T]) {
       .withRegion(Monitor) {
         x.ready.expect(true.B)
       }
-      .joinAndStep(getSourceClock)
+      .joinAndStep()
   }
 
   def enqueue(data: T): Unit = {
@@ -45,10 +32,10 @@ class DecoupledDriver[T <: Data](x: ReadyValidIO[T]) {
     fork
       .withRegion(Monitor) {
         while (!x.ready.peekBoolean()) {
-          getSourceClock.step(1)
+          step(1)
         }
       }
-      .joinAndStep(getSourceClock)
+      .joinAndStep()
   }
 
   def enqueueSeq(data: Seq[T]): Unit = {
@@ -64,23 +51,10 @@ class DecoupledDriver[T <: Data](x: ReadyValidIO[T]) {
     this
   }
 
-  def setSinkClock(clock: Clock): this.type = {
-    ClockResolutionUtils.setClock(DecoupledDriver.decoupledSinkKey, x, clock)
-    this
-  }
-
-  protected def getSinkClock: Clock = {
-    ClockResolutionUtils.getClock(
-      DecoupledDriver.decoupledSinkKey,
-      x,
-      x.valid.getSourceClock()
-    ) // TODO: validate against bits/valid sink clocks
-  }
-
   // NOTE: this doesn't happen in the Monitor phase, unlike public functions
   def waitForValid(): Unit = {
     while (!x.valid.peekBoolean()) {
-      getSinkClock.step(1)
+      step(1)
     }
   }
 
@@ -93,7 +67,7 @@ class DecoupledDriver[T <: Data](x: ReadyValidIO[T]) {
         x.valid.expect(true.B)
         x.bits.expect(data)
       }
-      .joinAndStep(getSinkClock)
+      .joinAndStep()
   }
 
   def expectDequeueNow(data: T): Unit = {
@@ -104,7 +78,7 @@ class DecoupledDriver[T <: Data](x: ReadyValidIO[T]) {
         x.valid.expect(true.B)
         x.bits.expect(data)
       }
-      .joinAndStep(getSinkClock)
+      .joinAndStep()
   }
 
   def expectDequeueSeq(data: Seq[T]): Unit = {
@@ -125,9 +99,9 @@ class DecoupledDriver[T <: Data](x: ReadyValidIO[T]) {
       x.valid.expect(false.B)
     }
   }
-}
 
-object DecoupledDriver {
-  protected val decoupledSourceKey = new Object()
-  protected val decoupledSinkKey = new Object()
+  @deprecated("You no longer need to set the clock explicitly.", since = "6.0.x")
+  def setSourceClock(clock: Clock): this.type = this
+  @deprecated("You no longer need to set the clock explicitly.", since = "6.0.x")
+  def setSinkClock(clock: Clock): this.type = this
 }