--- a/src/main/scala/chiseltest/RawTester.scala
+++ b/src/main/scala/chiseltest/RawTester.scala
@@ -5,15 +5,13 @@ package chiseltest
 import chiseltest.internal._
 import chisel3.Module
 import chiseltest.formal.Formal
-import chiseltest.internal.TestEnvInterface.addDefaultTargetDir
-import chiseltest.internal.TesterUtils.sanitizeFileName
 import firrtl2.AnnotationSeq
 
 /** Used to run simple tests that do not require a scalatest environment in order to run
   * @param testName
   *   This will be used to generate a working directory in ./test_run_dir
   */
-private class RawTester(testName: String) extends TestEnvInterface with HasTestName with Formal {
+private class RawTester(testName: String) extends HasTestName with Formal {
   // Provide test fixture data as part of 'global' context during test runs
   val topFileName = Some(testName)
 
@@ -23,9 +21,8 @@ private class RawTester(testName: String) extends TestEnvInterface with HasTestN
     chiselAnnos:   firrtl.AnnotationSeq = Seq()
   )(testFn:        T => Unit
   ): TestResult = {
-    batchedFailures.clear()
-    val newAnnos = addDefaultTargetDir(sanitizeFileName(testName), annotationSeq)
-    Context.runTest(this, () => dutGen, newAnnos, chiselAnnos, testFn)
+    val newAnnos = TesterUtils.addDefaultTargetDir(TesterUtils.sanitizeFileName(testName), annotationSeq)
+    Context.runTest(() => dutGen, newAnnos, chiselAnnos, testFn)
   }
 
   override def getTestName = testName