--- a/src/main/scala/chiseltest/exceptions.scala
+++ b/src/main/scala/chiseltest/exceptions.scala
@@ -8,6 +8,8 @@ class UnpeekableException(message: String) extends Exception(message)
 
 class ClockResolutionException(message: String) extends Exception(message)
 
+class FailedExpectException(val message: String, val failedCodeStackDepth: Int) extends Exception(message)
+
 class ThreadOrderDependentException(message: String) extends Exception(message)
 class TimeoutException(message: String)
     extends Exception(