--- a/src/main/scala/chiseltest/exceptions.scala
+++ b/src/main/scala/chiseltest/exceptions.scala
@@ -10,6 +10,8 @@ class ClockResolutionException(message: String) extends Exception(message)
 
 class FailedExpectException(val message: String, val failedCodeStackDepth: Int) extends Exception(message)
 
+class TemporalParadox(message: String) extends Exception(message)
+
 class ThreadOrderDependentException(val message: String, val failedCodeStackDepth: Int) extends Exception(message)
 class TimeoutException(message: String)
     extends Exception(